Implementado el listado de arnau

+++ src/main/java/com/example/astroterrassa/DAO/UsersRolesRepository.java
@@ -3,6 +3,6 @@ package com.example.astroterrassa.DAO;
 import com.example.astroterrassa.model.UsersRoles;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UsersRolesRepository extends JpaRepository<UsersRoles, Long>{
+public interface UsersRolesRepository extends JpaRepository<UsersRoles, Integer>{
 
 }
